Set up basic navigation Co-authored-by: Matthew Malone <m.malone@me.com>
