+++ components/Home/Home.js
@@ -1,0 +1,119 @@
+import React from 'react';
+import { StyleSheet, Text, View, Image, TextInput, Button, TouchableOpacity } from 'react-native';
+import * as ImagePicker from 'expo-image-picker';
+import * as Permissions from 'expo-permissions';
+import * as Camera from 'expo-camera';
+import * as Constants from 'expo-constants';
+import * as Font from 'expo-font';
+import { LinearGradient } from 'expo-linear-gradient';
+
+export default class App extends React.Component {
+  state = {
+    image: null
+  }
+
+  async componentDidMount() {
+    Font.loadAsync({
+      'major-mono-display': require('./assets/fonts/MajorMonoDisplay-Regular.ttf'),
+    });
+  }
+
+  render() {
+    let { image } = this.state
+    return (
+      <View style={styles.container}>
+        <Image source={require('./assets/PupDatesLogo.png')} style={styles.image}/>
+        <Text style={styles.title}>PupDates</Text>
+        <TextInput placeholder="Email" style={styles.input} />
+        <TextInput placeholder="Password" style={styles.input}/>
+        <LinearGradient
+          colors={['orange', '#c32525']}
+          style={styles.linearGradient}
+        >
+          <TouchableOpacity style={styles.button}><Text style={styles.buttonText}>Login</Text></TouchableOpacity>
+        </LinearGradient>
+        <Text>PupDates</Text>
+        <Button 
+          onPress={() => this.selectImg()}
+          title="Add Photo"
+        />
+        <Button 
+          onPress={() => this.takeImg()}
+          title="Camera"
+        />
+        {image &&
+            <Image source={{ uri: image }} style={{ width: 200, height: 200 }} />}
+      </View>
+    );
+  }
+
+selectImg() {
+  ImagePicker.requestCameraRollPermissionsAsync()
+    .then(response => {
+      if (response.granted === true) {
+        ImagePicker.launchImageLibraryAsync()
+          .then(response => this.setState({ image: response.uri }))
+      }
+    })
+  }
+
+takeImg() {
+  Camera.requestPermissionsAsync()
+    .then(response => {
+      if (response.granted === true) {
+        ImagePicker.launchCameraAsync()
+          .then(response => this.setState({ image: response.uri }))
+      }
+    })
+  }
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#fff',
+    alignItems: 'center',
+    marginTop: 90
+  },
+  image: {
+    height: 140,
+    width: 140,
+  },
+  title: {
+    fontSize: 50,
+    fontFamily: 'major-mono-display'
+  },
+  input: {
+    height: 30,
+    width: '70%',
+    borderColor: 'lightgrey',
+    borderRadius: 50,
+    // borderColor: 'rgba(33,33,33,0.81)',
+    borderWidth: 1.5,
+    padding: 10,
+    height: 60,
+    margin: 20,
+    alignItems: 'center'
+  },
+  linearGradient: {
+    width: '70%',
+    borderRadius: 25,
+    height: 50,
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+  button: {
+    // borderRadius: 50,
+    borderTopRightRadius: 20,
+    borderBottomRightRadius: 20,
+    borderBottomLeftRadius: 20,
+    borderTopLeftRadius: 20,
+    alignItems: 'center',
+    justifyContent: 'center',
+    color: 'black'
+  },
+  buttonText: {
+    color: '#fff',
+    fontSize: 25,
+  }
+});
